requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.4
psycopg2-binary==2.9.11
PyYAML==6.0
python-dotenv==0.21.0
python-engineio==4.12.3
python-socketio==5.15.0
gunicorn